requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.2
frozenlist==1.6.0
greenlet==3.2.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
mysql-connector-python==9.3.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
scheduler==0.8.8
SQLAlchemy==2.0.40
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
yarl==1.20.0
